[ADD] Two Scenes and connecting segues
