+++ PhotoFrame/PhotoFrame/Base.lproj/Main.storyboard
@@ -48,6 +48,7 @@
                                 <connections>
                                     <action selector="nextButtonTouchDown:" destination="2Vd-Ma-r6m" eventType="touchDown" id="kDI-a8-Rbs"/>
                                     <action selector="nextButtonTouched:" destination="2Vd-Ma-r6m" eventType="touchUpInside" id="kkn-OM-319"/>
+                                    <segue destination="sCE-6D-qet" kind="show" id="02u-Hm-XoB"/>
                                 </connections>
                             </button>
                             <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="vJY-hb-Xc1">
@@ -99,7 +100,7 @@
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="GyJ-yN-paG" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="-335" y="-7"/>
+            <point key="canvasLocation" x="1106" y="214"/>
         </scene>
         <!--My Tab Bar View Controller-->
         <scene sceneID="8xt-EX-VJc">
@@ -119,6 +120,61 @@
                 <placeholder placeholderIdentifier="IBFirstResponder" id="8z7-7B-Vfd" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="-1696" y="-308"/>
+        </scene>
+        <!--Inner View Controller-->
+        <scene sceneID="jxR-Ku-36s">
+            <objects>
+                <viewController id="sCE-6D-qet" customClass="InnerViewController" customModule="PhotoFrame" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="lVx-48-VLF">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="842"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="cYG-2A-NfV">
+                                <rect key="frame" x="182" y="467" width="57" height="35"/>
+                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                                <state key="normal" title="Button"/>
+                                <buttonConfiguration key="configuration" style="plain" title="다음">
+                                    <fontDescription key="titleFontDescription" name="MarkerFelt-Thin" family="Marker Felt" pointSize="19"/>
+                                </buttonConfiguration>
+                                <connections>
+                                    <segue destination="jmx-ho-bg3" kind="show" id="GVQ-fB-lGo"/>
+                                </connections>
+                            </button>
+                        </subviews>
+                        <viewLayoutGuide key="safeArea" id="ijc-8K-JSj"/>
+                        <color key="backgroundColor" red="0.93120485544204712" green="0.93367248773574829" blue="0.72977042198181152" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                    </view>
+                    <navigationItem key="navigationItem" id="cgh-Pr-YDr"/>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="E3x-87-5Ko" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="2128" y="-717"/>
+        </scene>
+        <!--Second Inner View Controller-->
+        <scene sceneID="eFX-bA-4y5">
+            <objects>
+                <viewController id="jmx-ho-bg3" customClass="SecondInnerViewController" customModule="PhotoFrame" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="Yre-x2-CUP">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="842"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="qr6-kJ-ifN">
+                                <rect key="frame" x="182" y="467" width="57" height="35"/>
+                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                                <state key="normal" title="Button"/>
+                                <buttonConfiguration key="configuration" style="plain" title="다음">
+                                    <fontDescription key="titleFontDescription" name="MarkerFelt-Thin" family="Marker Felt" pointSize="19"/>
+                                </buttonConfiguration>
+                            </button>
+                        </subviews>
+                        <viewLayoutGuide key="safeArea" id="yTr-Zq-U7v"/>
+                        <color key="backgroundColor" red="0.57122451066970825" green="0.91515570878982544" blue="0.62552034854888916" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                    </view>
+                    <navigationItem key="navigationItem" id="NFO-jt-Qgl"/>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="G1O-if-a9J" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="3055" y="-717"/>
         </scene>
     </scenes>
     <resources>
