--- PhotoFrame/PhotoFrame/SecondViewController.swift
@@ -1,29 +1,0 @@
-//
-//  SecondViewController.swift
-//  PhotoFrame
-//
-//  Created by Kai Kim on 2022/02/15.
-//
-
-import UIKit
-
-class SecondViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
