+++ PhotoFrame/PhotoFrame/Views/InnerViewController.swift
@@ -1,0 +1,18 @@
+//
+//  InnerViewController.swift
+//  PhotoFrame
+//
+//  Created by Kai Kim on 2022/02/17.
+//
+
+import UIKit
+
+class InnerViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+    }
+    
+
+}
