+++ PhotoFrame/PhotoFrame/Views/SecondInnerViewController.swift
@@ -1,0 +1,20 @@
+//
+//  SecondInnerViewController.swift
+//  PhotoFrame
+//
+//  Created by Kai Kim on 2022/02/17.
+//
+
+import UIKit
+
+class SecondInnerViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+    }
+    
+
+ 
+
+}
