+++ PhotoFrame/PhotoFrame/Views/SecondViewController.swift
@@ -1,0 +1,19 @@
+//
+//  SecondViewController.swift
+//  PhotoFrame
+//
+//  Created by Kai Kim on 2022/02/15.
+//
+
+import UIKit
+
+class SecondViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+    }
+    
+
+
+}
