+++ PhotoFrame/PhotoFrame/Views/myTabBarViewController.swift
@@ -1,0 +1,19 @@
+//
+//  myTabBarViewController.swift
+//  PhotoFrame
+//
+//  Created by Kai Kim on 2022/02/15.
+//
+
+import UIKit
+
+class myTabBarViewController: UITabBarController {
+
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        print(#file, #line, #function, #column)
+        
+    }
+   
+}
