--- PhotoFrame/PhotoFrame/myTabBarViewController.swift
@@ -1,27 +1,0 @@
-//
-//  myTabBarViewController.swift
-//  PhotoFrame
-//
-//  Created by Kai Kim on 2022/02/15.
-//
-
-import UIKit
-
-class myTabBarViewController: UITabBarController {
-
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        print(#file, #line, #function, #column)
-        // Do any additional setup after loading the view.
-        
-    }
-    /*
-    // MARK: - Navigation
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-}
